Refactor(member): 친구 목록, 친구 설정 컴포넌트 재구성
- member section에서 member type 선택 후
- 친구 목록, 친구 설정 section 분리해 컴포넌트 구성
